# docs/swagger.yaml
definitions:
  wallet.Err:
    properties:
      message:
        type: string
    type: object
  wallet.Wallet:
    properties:
      balance:
        example: 100
        type: number
      created_at:
        example: "2024-03-25T14:19:00.729237Z"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Create Card
        type: string
    type: object
host: localhost:1323
info:
  contact: {}
  description: Sophisticated Wallet API
  title: Wallet API
  version: "1.0"
paths:
  /api/v1/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all wallets
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Create wallet
      parameters:
      - description: wallet
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/wallet.Wallet'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wallet.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Create wallet
      tags:
      - wallet
    put:
      consumes:
      - application/json
      description: Update wallet
      parameters:
      - description: wallet
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/wallet.Wallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wallet.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Update wallet
      tags:
      - wallet
  /api/v1/wallets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete wallet
      parameters:
      - description: wallet id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Delete wallet
      tags:
      - wallet
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with wallet type
      parameters:
      - description: wallet type
        in: query
        name: wallet_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all users with wallet type
      tags:
      - wallet
  /users/{id}/wallets:
    get:
      consumes:
      - application/json
      description: Get wallet specific by user id
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get wallet specific by user id
      tags:
      - wallet
swagger: "2.0"